requirements.txt
wheel
flash-attn
transformers==4.48.1
datasets==3.1.0
accelerate==1.3.0
hf-transfer==0.1.
deepspeed==0.15.4
trl==0.14.0
vllm==0.7.0
bitsandbytes
peft
wandb
hydra-core